+++ src/js/validator/date.js
@@ -198,20 +198,20 @@
                 if (isNaN(Date.parse(minOption))) {
                     minOption = validator.getDynamicOption($field, minOption);
                 }
-                min = minOption instanceof Date ? minOption : this._parseDate(minOption, dateFormat, separator);
 
+                min       = minOption instanceof Date ? minOption : this._parseDate(minOption, dateFormat, separator);
                 // In order to avoid displaying a date string like "Mon Dec 08 2014 19:14:12 GMT+0000 (WET)"
-                minOption = minOption instanceof Date ? this._dateFormat(minOption, options.format) : minOption;
+                minOption = minOption instanceof Date ? this._formatDate(minOption, options.format) : minOption;
             }
 
             if (maxOption) {
                 if (isNaN(Date.parse(maxOption))) {
                     maxOption = validator.getDynamicOption($field, maxOption);
                 }
-                max = maxOption instanceof Date ? maxOption : this._parseDate(maxOption, dateFormat, separator);
 
+                max       = maxOption instanceof Date ? maxOption : this._parseDate(maxOption, dateFormat, separator);
                 // In order to avoid displaying a date string like "Mon Dec 08 2014 19:14:12 GMT+0000 (WET)"
-                maxOption = maxOption instanceof Date ? this._dateFormat(maxOption, options.format) : maxOption;
+                maxOption = maxOption instanceof Date ? this._formatDate(maxOption, options.format) : maxOption;
             }
 
             date = new Date(year, month -1, day, hours, minutes, seconds);
@@ -275,89 +275,90 @@
         },
 
         /**
-         * This function is based on the dateFormat function from the Date Format 1.2.3
-         * Credit to (c) 2007-2009 Steven Levithan <stevenlevithan.com>
-         * MIT license
-         * see http://blog.stevenlevithan.com/archives/date-time-format for the complete lib
+         * Format date
          *
-         * Return the date string formatted following the format provided as param
-         *
-         * @param {Date} date      The date object to format
-         * @param {String} format  The date format
-         * The format can be:
-         *   - date: Consist of DD, MM, YYYY parts which are separated by the separator option
-         *   - date and time:
-         *     The time can consist of h/hh/H/HH, M/MM, s/ss parts which are separated by :
-         * with
-         *      d	   Day of the month as digits; no leading zero for single-digit days.
-         *      dd	   Day of the month as digits; leading zero for single-digit days.
-         *      m	   Month as digits; no leading zero for single-digit months.
-         *      mm	   Month as digits; leading zero for single-digit months.
-         *      yy	   Year as last two digits; leading zero for years less than 10.
-         *      yyyy   Year represented by four digits.
-         *      h	   Hours; no leading zero for single-digit hours (12-hour clock).
-         *      hh	   Hours; leading zero for single-digit hours (12-hour clock).
-         *      H	   Hours; no leading zero for single-digit hours (24-hour clock).
-         *      HH	   Hours; leading zero for single-digit hours (24-hour clock).
-         *      M	   Minutes; no leading zero for single-digit minutes.
-         *             Uppercase M unlike CF timeFormat's m to avoid conflict with months.
-         *      MM	   Minutes; leading zero for single-digit minutes.
-         *             Uppercase MM unlike CF timeFormat's mm to avoid conflict with months.
-         *      s	   Seconds; no leading zero for single-digit seconds.
-         *      ss	   Seconds; leading zero for single-digit seconds.
+         * @param {Date} date The date object to format
+         * @param {String} format The date format
+         * The format can consist of the following tokens:
+         *      d       Day of the month without leading zeros (1 through 31)
+         *      dd      Day of the month with leading zeros (01 through 31)
+         *      m       Month without leading zeros (1 through 12)
+         *      mm      Month with leading zeros (01 through 12)
+         *      yy      Last two digits of year (for example: 14)
+         *      yyyy    Full four digits of year (for example: 2014)
+         *      h       Hours without leading zeros (1 through 12)
+         *      hh      Hours with leading zeros (01 through 12)
+         *      H       Hours without leading zeros (0 through 23)
+         *      HH      Hours with leading zeros (00 through 23)
+         *      M       Minutes without leading zeros (0 through 59)
+         *      MM      Minutes with leading zeros (00 through 59)
+         *      s       Seconds without leading zeros (0 through 59)
+         *      ss      Seconds with leading zeros (00 through 59)
          * @returns {String}
          */
-        _dateFormat: function(date, format) {
-            format = format.replace(/Y/g, "y")      // Replace the Y in year with LowerCase one
-                           .replace(/M/g, "m")      // Replace the M in month with LowerCase one
-                           .replace(/D/g, "d")      // Replace the D in day with LowerCase one
-                           .replace(/:m/g, ":M")    // Replace the minute character with UpperCase one
-                           .replace(/:mm/g, ":MM")
-                           .replace(/:S/, ":s")     // Replace the second character with LowerCase one
-                           .replace(/:SS/, ":ss");
+        _formatDate: function(date, format) {
+            format = format
+                        .replace(/Y/g, 'y')
+                        .replace(/M/g, 'm')
+                        .replace(/D/g, 'd')
+                        .replace(/:m/g, ':M')
+                        .replace(/:mm/g, ':MM')
+                        .replace(/:S/, ':s')
+                        .replace(/:SS/, ':ss');
 
-            var	token = /d{1,4}|m{1,4}|yy(?:yy)?|([HhMs])\1?|"[^"]*"|'[^']*'/g,
-                pad = function (val, len) {
-                    val = String(val);
-                    len = len || 2;
-                    while (val.length < len)
-                        val = "0" + val;
-                    return val;
-                };
+            var replacer = {
+                d: function(date) {
+                    return date.getDate();
+                },
+                dd: function(date) {
+                    var d = date.getDate();
+                    return (d < 10) ? '0' + d : d;
+                },
+                m: function(date) {
+                    return date.getMonth() + 1;
+                },
+                mm: function(date) {
+                    var m = date.getMonth() + 1;
+                    return m < 10 ? '0' + m : m;
+                },
+                yy: function(date) {
+                    return ('' + date.getFullYear()).substr(2);
+                },
+                yyyy: function(date) {
+                    return date.getFullYear();
+                },
+                h: function(date) {
+                    return date.getHours() % 12 || 12;
+                },
+                hh: function(date) {
+                    var h = date.getHours() % 12 || 12;
+                    return h < 10 ? '0' + h : h;
+                },
+                H: function(date) {
+                    return date.getHours();
+                },
+                HH: function(date) {
+                    var H = date.getHours();
+                    return H < 10 ? '0' + H : H;
+                },
+                M: function(date) {
+                    return date.getMinutes();
+                },
+                MM: function(date) {
+                    var M = date.getMinutes();
+                    return M < 10 ? '0' + M : M;
+                },
+                s: function(date) {
+                    return date.getSeconds();
+                },
+                ss: function(date) {
+                    var s = date.getSeconds();
+                    return s < 10 ? '0' + s : s;
+                }
+            };
 
-            // Passing date through Date applies Date.parse, if necessary
-            date = date ? new Date(date) : new Date;
-
-            if (isNaN(date))
-                throw new SyntaxError("invalid date");
-
-            format = String(format);
-
-            var	d = date["getDate"](),
-                m = date["getMonth"](),
-                y = date["getFullYear"](),
-                H = date["getHours"](),
-                M = date["getMinutes"](),
-                s = date["getSeconds"](),
-                flags = {
-                    d:    d,
-                    dd:   pad(d),
-                    m:    m + 1,
-                    mm:   pad(m + 1),
-                    yy:   String(y).slice(2),
-                    yyyy: y,
-                    h:    H % 12 || 12,
-                    hh:   pad(H % 12 || 12),
-                    H:    H,
-                    HH:   pad(H),
-                    M:    M,
-                    MM:   pad(M),
-                    s:    s,
-                    ss:   pad(s)
-                };
-
-            return format.replace(token, function ($0) {
-                return $0 in flags ? flags[$0] : $0.slice(1, $0.length - 1);
+            return format.replace(/d{1,4}|m{1,4}|yy(?:yy)?|([HhMs])\1?|"[^"]*"|'[^']*'/g, function(match) {
+                return replacer[match] ? replacer[match](date) : match.slice(1, match.length - 1);
             });
         }
     };